// src/interfaces/auth.interface.ts
export interface AuthResponse {
  access_token: string;
  user: {
    id: string;
    fullName: string;
    email: string;
    role: 'ADMIN' | 'AGENT' | 'CUSTOMER';
    profileImage?: string;
    phoneNumber?: string;
  };
}

export interface JwtPayload {
  sub: string;
  email: string;
  role: 'ADMIN' | 'AGENT' | 'CUSTOMER';
}